refactor(game): separate non-state classes to statistics package
