+++ src/main/java/chess/controller/ConsoleController.java
@@ -3,7 +3,7 @@ package chess.controller;
 import static chess.view.OutputView.print;
 
 import chess.domain.game.Game;
-import chess.domain.game.GameResult;
+import chess.domain.game.statistics.GameResult;
 import chess.domain.game.NewGame;
 import chess.domain.event.MoveCommand;
 import chess.dto.response.board.ConsoleBoardViewDto;
