+++ src/main/java/chess/domain/game/BlackTurn.java
@@ -1,9 +1,8 @@
 package chess.domain.game;
 
-import static chess.domain.board.piece.Color.BLACK;
-
 import chess.domain.board.Board;
 import chess.domain.board.piece.Color;
+import chess.domain.game.statistics.GameState;
 import java.util.Objects;
 
 final class BlackTurn extends Running {
@@ -14,7 +13,7 @@ final class BlackTurn extends Running {
 
     @Override
     public Color getCurrentTurnColor() {
-        return BLACK;
+        return Color.BLACK;
     }
 
     @Override
