+++ src/main/java/chess/domain/game/Game.java
@@ -1,6 +1,7 @@
 package chess.domain.game;
 
 import chess.domain.event.MoveCommand;
+import chess.domain.game.statistics.GameResult;
 import chess.dto.response.GameDto;
 import chess.dto.response.board.ConsoleBoardViewDto;
