+++ src/main/java/chess/domain/game/NewGame.java
@@ -2,6 +2,7 @@ package chess.domain.game;
 
 import chess.domain.board.Board;
 import chess.domain.event.MoveCommand;
+import chess.domain.game.statistics.GameResult;
 import chess.dto.response.GameDto;
 import chess.dto.response.board.ConsoleBoardViewDto;
 import chess.util.BoardMapGeneratorUtil;
