+++ src/main/java/chess/view/OutputView.java
@@ -4,7 +4,7 @@ import static chess.domain.board.piece.Color.BLACK;
 import static chess.domain.board.piece.Color.WHITE;
 
 import chess.domain.board.piece.Color;
-import chess.domain.game.GameResult;
+import chess.domain.game.statistics.GameResult;
 import chess.dto.response.board.ConsoleBoardViewDto;
 
 public class OutputView {
